Update frontend and review features

+++ schemas/review_schema.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel,Field
 
 class postReview(BaseModel):
     film_id:int
-    rating:int = Field(...,ge=1,le=5,description="Star rating from 1 to 5")
+    rating:float = Field(...,ge=1,le=5,description="Star rating from 1 to 5")
     review_text:str
